requirements.txt
streamlit==1.32.0
pandas==2.2.1
snowflake-connector-python==3.7.0
plotly==5.19.0
pillow==10.2.0
numpy==1.26.4
python-dotenv==1.0.1 